+++ remix/server/routes/api/[...].ts
@@ -22,6 +22,7 @@ const routeModules: Record<string, () => Promise<RouteModule>> = {
   'v1/auth/service-account': () => import('../../../app/routes/api/v1/auth/service-account/_service-account'),
   'v1/auth/verify-email': () => import('../../../app/routes/api/v1/auth/verify-email/_verify-email'),
   'v1/crypto': () => import('../../../app/routes/api/v1/crypto/_crypto'),
+  'v1/embed/things': () => import('../../../app/routes/api/v1/embed/things/_things'),
   'v1/health/frontend': () => import('../../../app/routes/api/v1/health/frontend/_frontend'),
   'v1/health/mongodb': () => import('../../../app/routes/api/v1/health/mongodb/_mongodb'),
   'v1/health/nitro': () => import('../../../app/routes/api/v1/health/nitro/_nitro'),